Refactored monolithic dev. code into a series of workflow package members
