+++ server/exercise/src/main/scala/com/eigengo/lift/exercise/classifiers/workflows/GestureWorkflows.scala
@@ -1,184 +1,10 @@
-package com.eigengo.lift.exercise
-
-package classifiers
+package com.eigengo.lift.exercise.classifiers.workflows
 
 import akka.stream.scaladsl._
-import akka.stream.stage.{TerminationDirective, PushStage, Context, Directive}
 import breeze.linalg.DenseMatrix
+import com.eigengo.lift.exercise.AccelerometerValue
 import com.eigengo.lift.exercise.classifiers.svm.{SVMClassifier, SVMModelParser}
 import com.typesafe.config.Config
-import scala.collection.mutable
-
-/**
- * Streaming stage that buffers events (the sample window). Once the buffer has filled, a callback  action is executed
- * for each new sample message that we receive.
- *
- * @param size   size of the internal buffer and so the sampling window size
- * @param action when buffer is full, action to be performed for each sample that we subsequently receive
- */
-class SamplingWindow[A] private (size: Int, action: List[A] => Unit) extends PushStage[A, A] {
-  require(size > 0)
-
-  private val buffer = mutable.Queue[A]()
-
-  override def onPush(elem: A, ctx: Context[A]): Directive = {
-    if (!ctx.isFinishing) {
-      if (buffer.length == size) {
-        // Buffer is full, so enable action callback
-        val emit = buffer.dequeue()
-        buffer.enqueue(elem)
-        action(buffer.toList)
-        // Allow old events to propagate downstream
-        ctx.push(emit)
-      } else {
-        // Buffer is not yet full, so keep consuming from our upstream
-        buffer.enqueue(elem)
-        ctx.pull()
-      }
-    } else {
-      // Streaming stage is shutting down, so we ensure that all buffer elements are flushed prior to finishing
-      if (buffer.isEmpty) {
-        // Buffer is empty, so we simply finish
-        ctx.finish()
-      } else if (buffer.length == 1) {
-        // One element in the buffer, so push it and finish
-        ctx.pushAndFinish(buffer.dequeue())
-      } else {
-        // Multiple elements are in the buffer, so reduce its size and eventually we will finish
-        ctx.push(buffer.dequeue())
-      }
-    }
-  }
-
-  override def onUpstreamFinish(ctx: Context[A]): TerminationDirective = {
-    ctx.absorbTermination()
-  }
-}
-
-object SamplingWindow {
-
-  def apply[A](size: Int)(action: List[A] => Unit): SamplingWindow[A] = {
-    new SamplingWindow(size, action)
-  }
-
-}
-
-/**
- * Streaming stage that buffers events (the sample window). Once the buffer has filled, a strategy function is applied
- * to determine the elements to be emitted. Emitted values are (typically) either a grouping of the elements making up
- * the sample or single sample elements.
- *
- * @param size     size of the internal buffer and so the sampling window size
- * @param strategy when buffer is full, sample window strategy that determines the elements to be emitted
- */
-class GroupBySample[A] private (size: Int, strategy: List[A] => GroupBySample.GroupValue[A]) extends PushStage[A, GroupBySample.GroupValue[A]] {
-  require(size > 0)
-
-  import GroupBySample._
-
-  private val buffer = mutable.Queue[A]()
-
-  private def applyStrategy(ctx: Context[GroupValue[A]]): Directive = {
-    strategy(buffer.toList) match {
-      case value @ BlobValue(sample) =>
-        // Aggregated sample to be emitted
-        buffer.drop(sample.length)
-        ctx.push(value)
-
-      case value: SingleValue[A] =>
-        // Single value sample is to be emitted
-        buffer.dequeue()
-        ctx.push(value)
-    }
-  }
-
-  override def onPush(elem: A, ctx: Context[GroupValue[A]]): Directive = {
-    if (!ctx.isFinishing) {
-      if (buffer.length == size) {
-        // Buffer is full, so apply our strategy to determine emit behaviour
-        applyStrategy(ctx)
-      } else {
-        // Buffer is not yet full, so keep consuming from our upstream
-        buffer.enqueue(elem)
-        ctx.pull()
-      }
-    } else {
-      // Streaming stage is shutting down, so we ensure that all buffer elements are flushed prior to finishing
-      if (buffer.isEmpty) {
-        // Buffer is empty, so we simply finish
-        ctx.finish()
-      } else if (buffer.length == 1) {
-        // One element in the buffer, so push it and finish
-        ctx.pushAndFinish(SingleValue(buffer.dequeue()))
-      } else if (buffer.length == size) {
-        // Buffer is full, so apply our strategy to determine emit behaviour
-        applyStrategy(ctx)
-      } else {
-        // Multiple elements are in the buffer (and it isn't full), so reduce its size and eventually we will finish
-        ctx.push(SingleValue(buffer.dequeue()))
-      }
-    }
-  }
-
-  override def onUpstreamFinish(ctx: Context[GroupValue[A]]): TerminationDirective = {
-    ctx.absorbTermination()
-  }
-}
-
-object GroupBySample {
-
-  /**
-   * Trait that allows data streams to be grouped into common values (c.f. `BlobValue`). Data values that can not be
-   * grouped are "marked" with `SingleValue`.
-   */
-  sealed trait GroupValue[A]
-  case class BlobValue[A](value: List[A]) extends GroupValue[A]
-  case class SingleValue[A](value: A) extends GroupValue[A]
-
-  def apply[A](size: Int)(strategy: List[A] => GroupValue[A]) = {
-    new GroupBySample[A](size, strategy)
-  }
-
-}
-
-/**
- * Flowgraph merge node that expects a message on each of its inputs. The collection of input messages is then outputted.
- * This merge node can be thought of as a generalised Zip or ZipWith node.
- *
- * @param size number of inputs that may be joined to this merge node
- */
-class ZipSet[A] private (size: Int) extends FlexiMerge[Set[A]] {
-  require(size >= 0)
-
-  import FlexiMerge._
-
-  val in = (0 until size).map { _ => createInputPort[A]() }.toVector
-
-  def createMergeLogic() = new MergeLogic[Set[A]] {
-    def initialState = State[ReadAllInputs](ReadAll(in)) { (ctx, _, inputs) =>
-      ctx.emit(in.flatMap(port => inputs.get[A](port)).toSet)
-
-      SameState[A]
-    }
-
-    def inputHandles(inputCount: Int) = {
-      require(inputCount == size, s"ZipSet must have $size connected inputs, was $inputCount")
-
-      in
-    }
-
-    override def initialCompletionHandling = eagerClose
-  }
-
-}
-
-object ZipSet {
-
-  def apply[A](size: Int) = {
-    new ZipSet[A](size)
-  }
-
-}
 
 object GestureTagging {
 
@@ -371,7 +197,7 @@ trait GestureWorkflows extends SVMClassifier {
    * @param inputLocations  locations that make up the input sensor network
    * @param outputLocations locations that make up the output sensor network
    */
-  class GestureClassificationWorkflow[L](inputLocations: Set[L], outputLocations: Set[L]) {
+  class GestureClassification[L](inputLocations: Set[L], outputLocations: Set[L]) {
     require(inputLocations.nonEmpty)
     require(outputLocations.nonEmpty)
 
@@ -418,8 +244,8 @@ trait GestureWorkflows extends SVMClassifier {
     }
   }
 
-  object GestureClassificationWorkflow {
-    def apply[L](inputLocations: Set[L], outputLocations: Set[L]) = new GestureClassificationWorkflow[L](inputLocations, outputLocations)
+  object GestureClassification {
+    def apply[L](inputLocations: Set[L], outputLocations: Set[L]) = new GestureClassification[L](inputLocations, outputLocations)
   }
 
 }
